visibility refactor
