--- px-checkout/src/main/java/com/mercadopago/android/px/internal/features/BankDealDetail.java
@@ -1,12 +1,0 @@
-package com.mercadopago.android.px.internal.features;
-
-import com.mercadopago.android.px.internal.base.MvpView;
-
-public interface BankDealDetail {
-
-    interface View extends MvpView {
-        void hideLogoName();
-
-        void hideLogo();
-    }
-}
